Refactoring model: working on base classes
Our data model is essentially build around getting data from gerrit and
streaming it through processing objects to get "materialized views" of
the data that we can show on the report.

So far we had an ad-hoc funciton-based implementation of this concept.

Now created some base classes to allow for a more modular and explicif
implementation of this concept.

Also added some unit tests and pytest to run them.

Signed-off-by: Barak Korren <bkorren@redhat.com>